+++ src/macros.rs
@@ -97,8 +97,8 @@
 ///
 /// assert_eq! {
 ///     vec![
-///         "import first",
 ///         "import default second",
+///         "import first",
 ///         "",
 ///         "first",
 ///         "default:second"
@@ -165,29 +165,25 @@ macro_rules! impl_lang {
         $(
             impl $crate::tokens::FormatInto<$lang> for $ty {
                 fn format_into(self, tokens: &mut $crate::Tokens<$lang>) {
-                    let item = <$lang as $crate::lang::Lang>::Item::from(self);
-                    tokens.append($crate::tokens::Item::Lang(Box::new(item)));
+                    tokens.append($crate::tokens::__lang_item::<$lang>(Box::new(self.into())));
                 }
             }
 
             impl<'a> $crate::tokens::FormatInto<$lang> for &'a $ty {
                 fn format_into(self, tokens: &mut $crate::Tokens<$lang>) {
-                    let item = <$lang as $crate::lang::Lang>::Item::from(self.clone());
-                    tokens.append($crate::tokens::Item::Lang(Box::new(item)));
+                    tokens.append($crate::tokens::__lang_item::<$lang>(Box::new(self.clone().into())));
                 }
             }
 
             impl $crate::tokens::Register<$lang> for $ty {
                 fn register(self, tokens: &mut $crate::Tokens<$lang>) {
-                    let item = <$lang as $crate::lang::Lang>::Item::from(self);
-                    tokens.append($crate::tokens::Item::Register(Box::new(item)));
+                    tokens.append($crate::tokens::__lang_item_register::<$lang>(Box::new(self.into())));
                 }
             }
 
             impl<'a> $crate::tokens::Register<$lang> for &'a $ty {
                 fn register(self, tokens: &mut $crate::Tokens<$lang>) {
-                    let item = <$lang as $crate::lang::Lang>::Item::from(self.clone());
-                    tokens.append($crate::tokens::Item::Register(Box::new(item)));
+                    tokens.append($crate::tokens::__lang_item_register::<$lang>(Box::new(self.clone().into())));
                 }
             }